pyproject.toml
[project]
name = "np-gh-actions"
version = "0.1.0"
description = ""
authors = [
    {name = "Christopher Mochizuki", email = "mochic808@gmail.com"},
]
dependencies = [
    "click>=8.1.6",
    "rich>=13.5.2",
    "pygithub>=1.59.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "python-dotenv>=1.0.0",
]
